updates 05_DEMO

+++ 05_DEMO/demo_check.sh
@@ -1,23 +1,22 @@
 #!/bin/bash
-set -e
+set -ex
 
-# sudo apt-get install -y r-base-core docker.io
+DOMAIN=$1
+CURL=$2
+
+apt-get update && apt-get install -y r-base-core docker.io
 
 # docker run -it --rm ymuski/curl-http3 curl -w "%{time_total}\n" -o /dev/null -s "https://demo.site --http3"
+# test domain response
+docker run -i --rm ymuski/curl-http3 $CURL -s https://$DOMAIN
+docker run -i --rm ymuski/curl-http3 ./httpstat.sh -Lv https://$DOMAIN
 
-for DOMAIN in {'eu.yurets.online','jp.yurets.online'}; do
-  if [ "$DOMAIN" = "eu.yurets.online" ]; then
-    CURL='curl --http3'
-  else
-    CURL='curl'
-  fi
-  # test domain response
-  docker run -it --rm ymuski/curl-http3 $CURL https://$DOMAIN
-  for i in {1..1000}; do
-    docker run -it --rm ymuski/curl-http3 $CURL -w "%{time_total}\n" -o /dev/null -s "https://${DOMAIN}" >> /tmp/${DOMAIN};
-  done;
-  echo -e "\n\n ${DOMAIN}";
-  cat /tmp/${DOMAIN} | Rscript -e 'stat=(scan("stdin")); print(quantile(stat, c(.50,.95,.99)));';
-  rm /tmp/${DOMAIN};
-echo -e "================== \n";
+[ -e /tmp/${DOMAIN} ] && rm /tmp/${DOMAIN};
+
+for i in {1..1000}; do
+  docker run -i --rm ymuski/curl-http3 $CURL -w "%{time_total}\n" -o /dev/null -s "https://${DOMAIN}" >> /tmp/${DOMAIN};
 done;
+echo -e "\n\n ${DOMAIN}";
+cat /tmp/${DOMAIN} | Rscript -e 'stat=(scan("stdin")); print(quantile(stat, c(.50,.95,.99)));' > /tmp/${DOMAIN}-results;
+
+echo -e "================== \n";
